Add array and object conditional fields support
